+++ packages/gamut/src/Markdown/__tests__/Markdown-test.tsx
@@ -2,6 +2,7 @@
 
 import { mount, ReactWrapper } from 'enzyme';
 import React from 'react';
+import { act } from 'react-dom/test-utils';
 
 import { Markdown } from '../index';
 
@@ -191,6 +192,23 @@ var test = true;
         <Markdown text={`<a href="http://google.com">google</a>`} />
       );
       expect(markdown.find('a[rel="noopener"]').length).toEqual(1);
+    });
+
+    it('Allows onClicks callbacks', () => {
+      const onClick = jest.fn();
+
+      const markdown = mount(
+        <Markdown
+          text={`<a data-testid="testLink" href="http://google.com">google</a>`}
+          onAnchorClick={onClick}
+        />
+      );
+
+      act(() => {
+        markdown.find(`a[href="http://google.com"]`).simulate('click');
+      });
+
+      expect(onClick).toHaveBeenCalledTimes(1);
     });
   });